refact: refactor error handling

+++ src/controllers/books/deleteBook.js
@@ -2,7 +2,11 @@ const BookRepository = require('../../repository/books');
 
 module.exports = {
     deleteBook: async (id) => {
-        const message = await BookRepository.removeBook(id);
-        return message
+        const book = await BookRepository.selectById(id);
+        if(!book) {
+            throw new Error('BOOK_DOES_NOT_EXIST');
+        };
+        const result = await BookRepository.removeBook(id);
+        return result;
     },
 };
